requirements.txt
fastapi
markdown2[all]
openai
pydantic_settings
python-multipart
uvicorn
